Move App.tsx, HomePage, NotFoundPage into /App
