+++ frontend/src/UnauthenticatedApp/index.tsx
@@ -5,8 +5,8 @@ import { useDispatch, useSelector } from "react-redux";
 import { Route, Switch } from "react-router-dom";
 import { toast } from "react-toastify";
 import api from "../api";
+import NotFoundPage from "../App/NotFoundPage";
 import { getRefreshToken } from "../localStorage";
-import NotFoundPage from "../components/NotFoundPage";
 import { setUser } from "../store/auth/actions";
 import { getUser } from "../store/auth/selectors";
 import ForgotPasswordPage from "./ForgotPasswordPage";
